requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
click==8.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1 
Flask-Mail==0.10.0

